requirements.txt
yfinance>=0.2.40
pandas>=2.0.0
numpy>=1.26.0
scipy>=1.11.0
requests>=2.31.0
beautifulsoup4>=4.12.0
openai>=1.30.0
alpaca-py>=0.20.0
streamlit>=1.35.0
plotly>=5.20.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
lxml>=4.9.0
ratelimit>=2.2.1
tenacity>=8.2.0
pdfkit>=1.0.0
